requirements.txt
flet==0.22.*
llama-index
faiss-cpu
langchain_google_genai
langchainhub
llama-index-llms-gemini
llama-index-embeddings-gemini
pytest-shutil
python-dotenv
fastapi
unicorn
redis
llama-index-packs-agents-coa
docx2txt
PyPDF2
pypdf
pymongo
pyperclip